feat(runtime): implement core evaluation loop and native function dispatch
Implement the central eval() recursive evaluator for the Pure runtime,
connecting the compiler's IR to the native function registry, variable
context, and object heap.

Key changes:

eval.rs — Complete evaluator implementation
- Evaluator<'model, H: EvalHooks> with generic hooks for zero-cost
  instrumentation (NoOpHooks compiles to nothing via monomorphization)
- Recursive eval() dispatches all 12 ExprKind variants: literals,
  variables, function calls, property access, lambdas, collections,
  enum values, type references
- EvalContext provides type-erased evaluator handle to native functions
- Lazy call stack: zero overhead on happy path, map_err + with_frame
  builds stack frames only when errors propagate

native.rs — EvalContextTrait and signature migration
- Added EvalContextTrait (object-safe) with eval_lambda(), context(),
  context_mut(), heap(), heap_mut()
- Updated NativeFunction::execute to accept &mut dyn EvalContextTrait
- Added defer_execution() for short-circuit semantics (mirrors Java's
  deferParameterExecution())
- Added NoOpEvalCtx test stub for simple native function tests
- Migrated all 5 native modules (arithmetic, boolean, comparison,
  collection, string) and their test call sites

value.rs — First-class lambda support
- Added Value::Lambda(LambdaClosure) variant
- LambdaClosure stores parameters, body, and captured variable bindings

types.rs / lower.rs — DateValue type alignment
- Changed DateValue field types from i32/u8/u32 to i16/i8/i32 to match
  jiff::civil and PureDate/StrictTime APIs, eliminating all runtime casts

error.rs — Lazy call stack builder
- Added PureException::with_frame(StackFrame) -> Self for the lazy
  call stack pattern used by the evaluator's recursive eval() chain

CLI fixes
- Renamed test command --verbose to --show-detail to avoid collision
  with global --verbose count flag
- Applied cargo clippy --fix auto-suggestions in parse/check/compile
- Updated test_stub_test_command integration test assertion

All quality gates pass: cargo fmt, cargo clippy --workspace (0 warnings),
cargo test --workspace (0 failures).

+++ legend-pure-rust/crates/pure/src/types.rs
@@ -149,44 +149,47 @@ pub struct Parameter {
 /// Parsed from the raw strings in `ast::StrictDateLiteral`, `ast::DateTimeLiteral`,
 /// and `ast::StrictTimeLiteral`. Avoids carrying raw strings through the
 /// semantic graph.
+///
+/// Field types match `jiff::civil` (and `PureDate`) so the evaluator needs
+/// zero casts.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum DateValue {
     /// Strict date: `%2024-01-15`.
     StrictDate {
-        /// Year.
-        year: i32,
+        /// Year (−9999..9999).
+        year: i16,
         /// Month (1–12).
-        month: u8,
+        month: i8,
         /// Day (1–31).
-        day: u8,
+        day: i8,
     },
     /// Date-time: `%2024-01-15T10:30:00`.
     DateTime {
-        /// Year.
-        year: i32,
+        /// Year (−9999..9999).
+        year: i16,
         /// Month (1–12).
-        month: u8,
+        month: i8,
         /// Day (1–31).
-        day: u8,
+        day: i8,
         /// Hour (0–23).
-        hour: u8,
+        hour: i8,
         /// Minute (0–59).
-        minute: u8,
+        minute: i8,
         /// Second (0–59).
-        second: u8,
-        /// Sub-second nanoseconds.
-        subsecond_nanos: u32,
+        second: i8,
+        /// Sub-second nanoseconds (`0–999_999_999`).
+        subsecond_nanos: i32,
     },
     /// Strict time: `%10:30:00`.
     StrictTime {
         /// Hour (0–23).
-        hour: u8,
+        hour: i8,
         /// Minute (0–59).
-        minute: u8,
+        minute: i8,
         /// Second (0–59).
-        second: u8,
-        /// Sub-second nanoseconds.
-        subsecond_nanos: u32,
+        second: i8,
+        /// Sub-second nanoseconds (`0–999_999_999`).
+        subsecond_nanos: i32,
     },
 }
